localfs: fix a goroutine/buffer leak
We weren't closing the pipe used to communicate with
the goroutine that does hashing, which meant that the
goroutine would get leaked.

Tested using pprof heap sampling and a local server.
Without the fix, the heap sample after isolating a test
looked like:

Showing top 10 nodes out of 27 (cum >= 557.26kB)
      flat  flat%   sum%        cum   cum%
   16384kB 88.61% 88.61%    16384kB 88.61%  encoding/json.(*Decoder).refill
 1548.03kB  8.37% 96.99%  1548.03kB  8.37%  io/ioutil.glob.func1
  557.26kB  3.01%   100%   557.26kB  3.01%  html.init

with the fix:

Showing top 10 nodes out of 21 (cum >= 16384kB)
      flat  flat%   sum%        cum   cum%
   16384kB 96.97% 96.97%    16384kB 96.97%  encoding/json.(*Decoder).refill
  512.50kB  3.03%   100%   512.50kB  3.03%  runtime.allocm